feat(app): Report and Handle error logs properly in download

+++ src/components/DownloadButton/index.tsx
@@ -1,18 +1,15 @@
 'use client';
 
 import React, { useEffect, useRef, useState } from 'react';
-import Link from 'next/link';
 
 import { CoreApi } from '@utils/core';
 import { getDownloadCounts } from '@utils/sponsor/get-count';
-import { bugReportTemplate } from '@utils/bug-report';
 
 import { DownloadCount } from './Counts';
 import { DownloadSponsor } from './Sponsor';
 import { DownloadSubButtons } from './SubButtons';
-import { ErrorSVG, ProcessingSVG } from './svgs';
-
-import { BUG_REPORT_ENDPOINT } from '@root/configs';
+import { DownloadError, ErrorLogs } from './Error';
+import { ProcessingSVG } from './svgs';
 
 import { Color } from 'bibata/app';
 import { Image } from 'bibata/core-api/types';
@@ -48,15 +45,28 @@ export const DownloadButton: React.FC<Props> = (props) => {
   const [lock, setLock] = useState<boolean>(false);
 
   const [loadingText, setLoadingText] = useState<string>('Preparing...');
-  const [errorText, setErrorText] = useState<string>('');
-  const [errorLogs, setErrorLogs] = useState<any>(null);
+  const [errorLogs, setErrorLogs] = useState<ErrorLogs>({ text: '' });
 
   const [showDropdown, setShowDropdown] = useState<boolean>(false);
   const buttonRef = useRef<HTMLButtonElement>(null);
   const dropdownRef = useRef<HTMLDivElement>(null);
 
-  const print_error = async (e: any) => {
+  const printError = async (e: any) => {
     if (process.env.NODE_ENV === 'development') console.error(e);
+  };
+
+  const updateErrorLogs = async (opt: {
+    text: string;
+    key: string;
+    error: any;
+  }) => {
+    setErrorLogs({
+      ...errorLogs,
+      id: api.jwt?.id,
+      role: api.jwt?.role,
+      text: opt.text,
+      [opt.key]: opt.error
+    });
   };
 
   const processImages = async (imgs: Image[], options: ProcessOptions) => {
@@ -75,6 +85,11 @@ export const DownloadButton: React.FC<Props> = (props) => {
 
       const upload = await api.uploadImages(formData);
       if (upload?.error) {
+        updateErrorLogs({
+          key: 'upload',
+          error: upload.error,
+          text: 'Oops.. Processing Falied! Try Again.'
+        });
         return upload;
       }
     }
@@ -86,7 +101,7 @@ export const DownloadButton: React.FC<Props> = (props) => {
     document.body.appendChild(link);
     link.click();
     link.parentNode?.removeChild(link);
-    setErrorText('');
+    setErrorLogs({ text: '' });
   };
 
   const storeToDB = async (platform: string) => {
@@ -107,20 +122,25 @@ export const DownloadButton: React.FC<Props> = (props) => {
         })
       });
     } catch (e) {
-      setErrorLogs(e);
-      print_error(e);
+      updateErrorLogs({
+        key: 'count',
+        error: e,
+        text: 'Unexpected Internal Error.'
+      });
+      printError(e);
     }
   };
 
   const handleDownload = async (platform: Platform) => {
     setLoading(true);
     setLock(true);
-    setErrorLogs(null);
+
+    setErrorLogs({ text: '' });
 
     api.refreshSession(tokenRef.current);
     const { count, total } = await getDownloadCounts(api.jwt?.token);
     if ((total && count >= total) || (count === 0 && total === 0)) {
-      setErrorText('Download Limit Exceeded.');
+      setErrorLogs({ text: 'Download Limit Exceeded.' });
     } else {
       const downloadUrl = api.downloadUrl(platform, name);
 
@@ -137,21 +157,24 @@ export const DownloadButton: React.FC<Props> = (props) => {
         });
 
         if (upload?.error) {
-          print_error(upload.error);
-          setErrorLogs(upload.error);
-          setErrorText('Oops.. Processing Falied! Try Again.');
+          printError(upload.error);
           await api.refreshSession();
         } else {
           setLoadingText(
-            `Packaging ${platform == 'win' ? 'Win Cursors' : 'XCursors'} ...`
+            `Packaging ${platform === 'win' ? 'Win Cursors' : 'XCursors'} ...`
           );
 
           const download = await api.downloadable(platform, name);
 
           if (download?.error) {
-            print_error(download.error);
-            setErrorLogs(download.error);
-            setErrorText('Oops.. Packaging Failed! Try Again.');
+            printError(download.error);
+
+            updateErrorLogs({
+              key: 'download',
+              error: download.error,
+              text: 'Oops.. Packaging Failed! Try Again.'
+            });
+
             await api.refreshSession();
           } else {
             await storeToDB(platform);
@@ -185,12 +208,12 @@ export const DownloadButton: React.FC<Props> = (props) => {
   }, []);
 
   useEffect(() => {
-    if (!lock) {
+    if (!lock && props.config !== configRef.current) {
       api.deleteSession();
       tokenRef.current = props.token;
       configRef.current = props.config;
     }
-  }, [lock, props.token]); // eslint-disable-line react-hooks/exhaustive-deps
+  }, [lock, props.token, props.config]); // eslint-disable-line react-hooks/exhaustive-deps
 
   const busy = loading || props.disabled;
 
@@ -223,29 +246,10 @@ export const DownloadButton: React.FC<Props> = (props) => {
                 </div>
               ) : (
                 <>
-                  {errorText && (
-                    <div className='flex p-6 justify-center items-center fill-red-300 text-red-300'>
-                      <div className='mr-1 mt-1 h-7 w-7'>
-                        <ErrorSVG />
-                      </div>
-                      <p className='font-bold'>{errorText}</p>
-                    </div>
-                  )}
-
-                  {errorLogs && (
-                    <div className='-mt-4 mb-3 flex justify-center'>
-                      <Link
-                        className='bg-red-500 hover:bg-red-400 px-4 py-2 rounded-xl '
-                        target='_blank'
-                        href={BUG_REPORT_ENDPOINT(
-                          errorText,
-                          bugReportTemplate(errorText, errorLogs)
-                        )}>
-                        Report Bug
-                      </Link>
-                    </div>
-                  )}
-
+                  <DownloadError
+                    logs={errorLogs}
+                    onClick={() => setErrorLogs({ text: '' })}
+                  />
                   <DownloadSubButtons
                     disabled={props.disabled || loading}
                     onClick={(p) => handleDownload(p)}
